# perf(all): Optimize modulo operations for ~5% faster lookups and ~10% faster generator