## Summary
This PR optimizes the two hot-path modulo operations in phf map/set lookups and in phf_generator. It replaces thm with a faster, branchless algorithm based on the paper ["Faster Remainder by Direct Computation" by Lemire, Kaser, and Kurz](https://arxiv.org/abs/1902.01961).

The exhaustive search in the generator is slow, this changes replaces the expensive modulo on the hot path with 2 multiplications and one shift. Same can be said for runtime lookups.

## Performance Benchmarks
All benchmarks were run on an x86-64 CPU and are included in the benches/ directory for reproduction.

### Compile-Time: phf_generator Speed

| Input Size | Baseline (Standard `%`) | Experiment (Fast Modulo) | Improvement |
| :--- | :--- | :--- | :--- |
| 10 | 1.34 µs | 1.19 µs | **+11.2%** |
| 100 | 19.52 µs | 18.99 µs | **+2.7%** |
| 1,000 | 903.15 µs | 788.39 µs | **+12.7%** |
| 10,000 | 7.19 ms | 6.41 ms | **+10.8%** |
| 100,000 | 90.91 ms | 82.23 ms | **+9.5%** |
| 500,000 | 641.56 ms | 578.43 ms | **+9.8%** |
| 1,000,000 | 1.92 s | 1.75 s | **+8.9%** |

### Runtime: phf Lookup Speed

| Input Size | Baseline (Standard `%`) | Experiment (Fast Modulo) | Improvement |
| :--- | :--- | :--- | :--- |
| 10 | 103.88 ns | 100.50 ns | **~3.3%** |
| 100 | 1.01 µs | 0.99 µs | **~1.5%** |
| 1,000 | 10.00 µs | 9.82 µs | **~1.8%** |
| 10,000 | 104.02 µs | 101.89 µs | **~2.0%** |
| 100,000 | 1.24 ms | 1.18 ms | **~5.1%** |
| 500,000 | 9.29 ms | 8.87 ms | **~4.6%** |
| 1,000,000 | 22.83 ms | 21.69 ms | **~5.0%** |

## Alternative Approaches Considered

This is based on multiplication and a right-shift. But with this approach the generator was unable to find a collision-free displacement set, failing to build a map even for as few as 10 entries.

Link: [https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/](https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/)
```rust
fn index(hash: u32, range_len: u32) -> u32 {
    ((hash as u64 * range_len as u64) >> 32) as u32
}